recursively grab files from directories
